+++ flutter_dex/lib/components/bottomNavBar.dart
